avg rating theo product

+++ src/main/java/com/ndt/be_stepupsneaker/core/client/dto/request/review/ClientReviewRequest.java
@@ -1,13 +1,6 @@
 package com.ndt.be_stepupsneaker.core.client.dto.request.review;
 
 import com.ndt.be_stepupsneaker.core.common.base.PageableRequest;
-import com.ndt.be_stepupsneaker.entity.customer.Customer;
-import com.ndt.be_stepupsneaker.entity.product.ProductDetail;
-import com.ndt.be_stepupsneaker.infrastructure.constant.EntityProperties;
-import jakarta.persistence.Column;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -21,7 +14,7 @@ public class ClientReviewRequest extends PageableRequest {
 
     private String comment;
 
-    private int rating;
+    private double rating;
 
     private String urlImage;
